+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fix(tui): on track change, dont select "current track" playlist item if the old "current track" was not selected.
 - Fix(tui): re-select the (approximately) same playlist item after a shuffle.
 - Fix(server): with `rusty-soundtouch` on `rusty` backend, dont take initial samples until necessary.
+- Fix(server): on rusty backend, always decode and use `f32` samples. (instead of `i16`)
 - Fix: dont consider non-path Urls (podcasts, radio) for removal after a delete.
 
 ### [V0.10.0]